WIP: Parse MIME-Version in parts too
RFC2045 is unclear - Section 4 only states that MIME-Version is not
required for message parts, but it doesn't prohibit it. The grammar in
Appendix A in stating that fields which don't begin Content- can have no
meaning appears to have overlooked the possibility of this field being
written. Given that it is written by various MUAs, it seems better to
parse it rather than ignore it completely.

One option may be to parse it, but also put it in the fields list?